Updated controllers and model s to implement user access control and relations with users.

+++ app/Http/Controllers/EmployeeController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use Auth;
 use App\Employee;
 use Illuminate\Http\Request;
 
@@ -12,6 +13,10 @@ class EmployeeController extends Controller
      */
     public function index()
     {
+        if(Auth::user()->role != 'admin') {
+            return redirect('order');
+        }
+
         // Retrieve a list of all employees from db
         $employees = Employee::all();
 
@@ -25,6 +30,10 @@ class EmployeeController extends Controller
      */
     public function edit($id)
     {
+        if(Auth::user()->role != 'admin') {
+            return redirect('order');
+        }
+
         // Find the employee with id : $id
         $employee = Employee::findOrFail($id);
 
@@ -34,8 +43,13 @@ class EmployeeController extends Controller
 
     public function update(Request $request, $id)
     {
+        if(Auth::user()->role != 'admin') {
+            return redirect('order');
+        }
+
         // Find the employee with id : $id and update its properies with the posted data
         $employee = Employee::findOrFail($id);
+        // $employee->user()->first()->update([$employee])
         $employee->update($request->all());
 
         // Redirect borwser to the /employee page (employees index page)
@@ -44,6 +58,10 @@ class EmployeeController extends Controller
 
     public function delete($id)
     {
+        if(Auth::user()->role != 'admin') {
+            return redirect('order');
+        }
+
         // Delete the employee with id: $id and redirect to employees index page
         Employee::destroy($id);
         return redirect('employee');
@@ -51,12 +69,20 @@ class EmployeeController extends Controller
 
     public function add()
     {
+        if(Auth::user()->role != 'admin') {
+            return redirect('order');
+        }
+
         // Show the employees/add.blade.php form for adding a new employee
         return view('employees.add');
     }
 
     public function create(Request $request)
     {
+        if(Auth::user()->role != 'admin') {
+            return redirect('order');
+        }
+
         // Create a new employee using the posted data and redirect to employees index page
         $employee = Employee::create($request->all());
         return redirect('employee');
